* commit 'bacdfc0ed9a0882ee3e9ee889a8fd4cabb9a71f3':
  Fix compiling warnings

+++ vdc.c
@@ -57,7 +57,7 @@ int main(int argc, char **argv) {
 static int do_cmd(int sock, int argc, char **argv) {
     char final_cmd[255] = "0 "; /* 0 is a (now required) sequence number */
     int i;
-    int ret;
+    size_t ret;
 
     for (i = 1; i < argc; i++) {
         char *cmp;
